refactor: mentor feedback
- add index.js for each component
- set min and max values for minutes/seconds, make them required
