+++ src/components/NewTaskForm/index.js
@@ -1,0 +1,3 @@
+import NewTaskForm from './NewTaskForm.jsx'
+
+export { NewTaskForm }
