+++ src/components/Task/index.js
@@ -1,0 +1,3 @@
+import Task from './Task.jsx'
+
+export { Task }
