+++ src/components/TaskFilter/index.js
@@ -1,0 +1,3 @@
+import TaskFilter from './TaskFilter.jsx'
+
+export { TaskFilter }
