+++ src/components/TaskList/index.js
@@ -1,0 +1,3 @@
+import TaskList from './TaskList.jsx'
+
+export { TaskList }
